Move xjson / xyaml to utils instead of prjxray.
